Fix 404 Error on XHR
Added code that catches a 404 status code when attempting to load
content that doesn't exist. Retrieves the 404 error page and shows it to
the user. This fixes issue #148 where pressing the back button after
visting a 404 page causes the previous page's content not to load
forcing the user to reload the web page.